% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alliances_regime.R
\docType{data}
\name{alliances_regime}
\alias{alliances_regime}
\title{Regime types and defense alliances}
\format{
\subsection{\code{alliances_regime}}{

A data frame with 18915 rows and 4 columns:

\describe{
\item{country_1}{The first country in the dyad}
\item{country_2}{The second country in the dyad}
\item{defense_agreement}{A binary variable indicating whether the two countries held a mutual defense treaty in 2017.}
\item{regime_dyad}{A binary variable indicating whether the two countries shared similar regime types ("Joint") or dissimilar regimes ("Mixed") in 2017.}
}
}
}
\usage{
alliances_regime
}
\description{
Information on all country-dyad regime type and defense alliance status in 2017.
}
\details{
Polity 2 and ATOP.
}
\keyword{datasets}
