% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_geo_spread.R
\docType{data}
\name{mid_geo_spread}
\alias{mid_geo_spread}
\title{The Geographic Spread of Militarized Disputes}
\format{
\subsection{\code{mid_geo_spread}}{

A data frame with 296 observations and 17 variables:

\describe{
\item{dispnum}{Unique dispute number from UCDP/PRIO Armed Conflict Data Set}
\item{dyadid}{Unique identification for the disputing dyad}
\item{ccode1}{Country one's COW code}
\item{ccode2}{Country two's COW code}
\item{year}{Year}
\item{log_radius_area}{The geographic spread of the MID (measured as a circular area measured in squared kms, the radius of which is the distance from the MID’s geometric center location to the furthest outlying incident)}
\item{territory}{At least one participant in the MID considers territory to be the primary issue of the conflict}
\item{host_resource}{The host state has a valuable resource endowment}
\item{host_mountain}{Extent of mountainous terrain in the host state}
\item{host_forest}{Extent of forest-covered terrain in the host state}
\item{ocean}{Location of initial incident is at sea}
\item{vital_border}{The ‘vitalness’ of a shared border}
\item{joint_size}{The joint size of the territories of the two states (logged)}
\item{peace_years}{The length of time since the two states previously experienced a MID}
\item{incidents}{Number of incidents in the MID}
\item{duration}{Temporal duration of the MID (logged)}
\item{hostility}{Highest hostility level experienced during the MID}
}
}
}
\source{
https://files.prio.org/Journals/JPR/2006/43/5/Replication\%20Braithwaite.zip
}
\usage{
mid_geo_spread
}
\description{
Data produced by Braithwaite (2006) on the relationship between the geographic spread of Militarized Interstate Disputes (MIDs) and geographic features.
}
\keyword{datasets}
