% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nes.R
\docType{data}
\name{nes}
\alias{nes}
\title{American National Election Studies 2020 Time Series}
\format{
\subsection{\code{nes}}{

A data frame with 8280 rows and 5 columns:

\describe{
\item{respondent_id}{Each respondent's identification number}
\item{attention_to_politics}{How often do you pay attention to what’s going on in government and politics?}
\item{education}{What is the highest level of school you have completed or the highest degree you have received?}
\item{therm_biden}{On a scale from 0 to 100, how would you rate Joe Biden?}
\item{therm_trump}{On a scale from 0 to 100, how would you rate Donald Trump?}
}
}
}
\source{
https://electionstudies.org/data-center/2020-time-series-study/
}
\usage{
nes
}
\description{
The ANES Time Series Study collects data on US voters' views on a range of political issues prior to and following a US Presidential Election.
}
\details{
American National Election Studies. 2021. ANES 2020 Time Series Study Full Release. February 10, 2022 version. www.electionstudies.org
}
\keyword{datasets}
